Arreglo menor

+++ Social.ME/src/main/kotlin/com/example/socialme/controller/ActividadController.kt
@@ -77,7 +77,7 @@ class ActividadController {
     @GetMapping("/verActividadesPublicasEnZona/{distanciaKm}/{username}")
     fun verActividadesPublicasEnZona(
         httpRequest: HttpServletRequest,
-        @PathVariable distanciaKm: Int,
+        @PathVariable distanciaKm: Float,
         @PathVariable username: String
     ): ResponseEntity<List<ActividadDTO>>{
         return  ResponseEntity(actividadService.verActividadesPublicas(distanciaKm, username), HttpStatus.OK)
